refactor(dashboardsettings): separate out linksettings components and concerns
