+++ public/app/features/dashboard/components/LinksSettings/index.tsx
@@ -1,0 +1,3 @@
+export { LinkSettingsEdit } from './LinkSettingsEdit';
+export { LinkSettingsHeader } from './LinkSettingsHeader';
+export { LinkSettingsList } from './LinkSettingsList';
